// src/app/modules/simple-angular/modules/sn-carousel/sn-carousel.interface.ts
export interface SnCarousel {
	auto?: boolean
	timer?: number
	reverse?: boolean
	pause_on_mouseover?: boolean
	active_item?: number
	item_gap?: string
	bullet_enabled?: boolean
	bullet_size?: string
	bullet_gap?: string
	bullet_bottom?: string
	bullet_top?: string
	bullet_bg_color?: string
	bullet_active_bg_color?: string
	arrow_enabled?: boolean
	arrow_disabled_on_reach_end?: boolean
	arrow_size?: string
	arrow_icon_size?: string
	arrow_bg_color?: string
	arrow_color?: string
	arrow_container?: string
}